split out unit tests for endpoints into own modules, add ipdb to pytest, add requirements
