+++ workshop-registration/requirements-dev.txt
@@ -1,0 +1,3 @@
+python-dotenv
+behave
+ipdb
