+++ workshop-registration/requirements.txt
@@ -1,2 +1,4 @@
-python-dotenv
-behave
+PyYAML
+dotenv
+pandas
+requests
